mk: Improve STRIP_DEBUG.
The first issue was that on platforms where strip was set to a noop the
command would succeed but the subsequent mv would fail.  Improve this by
actually checking for the output file.

The second issue is that not all strip commands support -g, so we need
an explicit gstrip tool for GNU strip (or equivalents, the FreeBSD
"elftoolchain" supports -g too).

While here simplify and tidy the install-strip-debug target, as well as
improve performance by skipping symlinks.

+++ mk/install/install.mk
@@ -370,16 +370,17 @@ install-ctf: plist
 .PHONY: install-strip-debug
 install-strip-debug: plist
 	@${STEP_MSG} "Automatic stripping of debug information"
-	${RUN}${CAT} ${_PLIST_NOKEYWORDS} \
-	| ${SED} -e 's|^|${DESTDIR}${PREFIX}/|' \
-	| while read f; do \
-		tmp_f="$${f}.XXX"; \
-		if ${STRIP} -g -o "$${tmp_f}" "$${f}" 2> /dev/null; then \
-			[ ! -f "$${f}" ] || \
-			    ${MV} "$${tmp_f}" "$${f}"; \
-		else \
-			${RM} -f "$${tmp_f}"; \
-		fi \
+	${RUN}${CAT} ${_PLIST_NOKEYWORDS}				\
+	| ${SED} -e 's|^|${DESTDIR}${PREFIX}/|'				\
+	| while read f; do						\
+		[ ! -h "$${f}" ] || continue;				\
+		tmp_f="$${f}.XXX";					\
+		if ${STRIP_DBG} -o "$${tmp_f}" "$${f}" 2>/dev/null; then \
+			if [ -f "$${tmp_f}" -a -f "$${f}" ]; then	\
+				${MV} "$${tmp_f}" "$${f}";		\
+			fi;						\
+		fi;							\
+		${RM} -f "$${tmp_f}";					\
 	done
 
 ######################################################################
